+++ playwright-e2e/tests/smoke-test.spec.ts
@@ -1,5 +1,5 @@
 import { test } from "../fixtures/create-fixture";
-import { newSwanseaLocalAuthorityUserOne } from "../settings/userCredentials";
+import { newSwanseaLocalAuthorityUserOne } from "../settings/user-credentials";
 
 test("Smoke Test @smoke-test", async ({
   signInPage,
